requirements.txt
requests>=2.24.0
numpy>=1.19.2
python-twitter>=3.5
mock>=4.0.2
